+++ actions/ReportsPercentiles.php
@@ -3,18 +3,19 @@
 namespace Modules\MoreReporting\Actions;
 
 use API;
-use CAbsoluteTimeParser;
 use CArrayHelper;
 use CController;
 use CControllerResponseData;
 use CControllerResponseFatal;
 use CProfile;
+use CRangeTimeParser;
 use CTimezoneHelper;
 use DateTimeZone;
 
 use Modules\MoreReporting\Includes\NativeGraph;
 use Modules\MoreReporting\Includes\ReportStorage;
 use Modules\MoreReporting\Includes\Reports\ItemPercentilesReport;
+use Modules\MoreReporting\Includes\TimePresets;
 
 class ReportsPercentiles extends CController {
 
@@ -28,6 +29,7 @@ class ReportsPercentiles extends CController {
 		$fields = [
 			'reportid' =>			'id',
 			'filter_groupids' =>	'array_db hosts_groups.groupid',
+			'filter_hostids' =>	'array_db hosts.hostid',
 			'filter_pattern' =>	'string',
 			'filter_date_from' =>	'string',
 			'filter_date_to' =>	'string',
@@ -51,39 +53,54 @@ class ReportsPercentiles extends CController {
 	protected function doAction(): void {
 		$definition = $this->hasInput('reportid') ? ReportStorage::get((int) $this->getInput('reportid')) : null;
 
-		if ($this->hasInput('filter_set')) {
-			CProfile::updateArray(self::PROFILE_PREFIX.'groupids', $this->getInput('filter_groupids', []),
-				PROFILE_TYPE_ID
-			);
-			CProfile::update(self::PROFILE_PREFIX.'pattern', $this->getInput('filter_pattern', ''), PROFILE_TYPE_STR);
-			CProfile::update(self::PROFILE_PREFIX.'date_from', $this->getInput('filter_date_from', ''),
-				PROFILE_TYPE_STR
-			);
-			CProfile::update(self::PROFILE_PREFIX.'date_to', $this->getInput('filter_date_to', ''), PROFILE_TYPE_STR);
-		}
-		elseif ($this->hasInput('filter_rst')) {
-			CProfile::deleteIdx(self::PROFILE_PREFIX.'groupids');
-			CProfile::delete(self::PROFILE_PREFIX.'pattern');
-			CProfile::delete(self::PROFILE_PREFIX.'date_from');
-			CProfile::delete(self::PROFILE_PREFIX.'date_to');
-		}
+		if ($definition !== null) {
+			$config = $definition['config'];
 
-		$filter = $definition !== null
-			? [
-				'groupids' => $definition['config']['groupids'] ?? [],
-				'pattern' => $definition['config']['pattern'] ?? '',
-				'date_from' => CProfile::get(self::PROFILE_PREFIX.'date_from', date(ZBX_DATE_TIME, time() - SEC_PER_DAY)),
-				'date_to' => CProfile::get(self::PROFILE_PREFIX.'date_to', date(ZBX_DATE_TIME, time()))
-			]
-			: [
-				'groupids' => CProfile::getArray(self::PROFILE_PREFIX.'groupids', []),
-				'pattern' => CProfile::get(self::PROFILE_PREFIX.'pattern', ''),
-				'date_from' => CProfile::get(self::PROFILE_PREFIX.'date_from', date(ZBX_DATE_TIME, time() - SEC_PER_DAY)),
-				'date_to' => CProfile::get(self::PROFILE_PREFIX.'date_to', date(ZBX_DATE_TIME, time()))
+			$filter = [
+				'groupids' => $config['groupids'] ?? [],
+				'hostids' => $config['hostids'] ?? [],
+				'pattern' => $config['pattern'] ?? '',
+				'date_from' => $this->getInput('filter_date_from', $config['period']['from'] ?? 'now-7d'),
+				'date_to' => $this->getInput('filter_date_to', $config['period']['to'] ?? 'now')
 			];
+		}
+		else {
+			if ($this->hasInput('filter_set')) {
+				CProfile::updateArray(self::PROFILE_PREFIX.'groupids', $this->getInput('filter_groupids', []),
+					PROFILE_TYPE_ID
+				);
+				CProfile::updateArray(self::PROFILE_PREFIX.'hostids', $this->getInput('filter_hostids', []),
+					PROFILE_TYPE_ID
+				);
+				CProfile::update(self::PROFILE_PREFIX.'pattern', $this->getInput('filter_pattern', ''),
+					PROFILE_TYPE_STR
+				);
+				CProfile::update(self::PROFILE_PREFIX.'date_from', $this->getInput('filter_date_from', ''),
+					PROFILE_TYPE_STR
+				);
+				CProfile::update(self::PROFILE_PREFIX.'date_to', $this->getInput('filter_date_to', ''),
+					PROFILE_TYPE_STR
+				);
+			}
+			elseif ($this->hasInput('filter_rst')) {
+				CProfile::deleteIdx(self::PROFILE_PREFIX.'groupids');
+				CProfile::deleteIdx(self::PROFILE_PREFIX.'hostids');
+				CProfile::delete(self::PROFILE_PREFIX.'pattern');
+				CProfile::delete(self::PROFILE_PREFIX.'date_from');
+				CProfile::delete(self::PROFILE_PREFIX.'date_to');
+			}
+
+			$filter = [
+				'groupids' => CProfile::getArray(self::PROFILE_PREFIX.'groupids', []),
+				'hostids' => CProfile::getArray(self::PROFILE_PREFIX.'hostids', []),
+				'pattern' => CProfile::get(self::PROFILE_PREFIX.'pattern', ''),
+				'date_from' => CProfile::get(self::PROFILE_PREFIX.'date_from', 'now-7d'),
+				'date_to' => CProfile::get(self::PROFILE_PREFIX.'date_to', 'now')
+			];
+		}
 
 		$timezone = new DateTimeZone(CTimezoneHelper::getSystemTimezone());
-		$time_parser = new CAbsoluteTimeParser();
+		$time_parser = new CRangeTimeParser();
 
 		$time_parser->parse($filter['date_from']);
 		$time_from = $time_parser->getDateTime(true, $timezone)->getTimestamp();
@@ -98,10 +115,18 @@ class ReportsPercentiles extends CController {
 			]), ['groupid' => 'id'])
 			: [];
 
+		$hosts = $filter['hostids']
+			? CArrayHelper::renameObjectsKeys(API::Host()->get([
+				'output' => ['hostid', 'name'],
+				'hostids' => $filter['hostids']
+			]), ['hostid' => 'id'])
+			: [];
+
 		$report = new ItemPercentilesReport();
 
 		$raw_data = $report->getData([
 			'groupids' => $filter['groupids'],
+			'hostids' => $filter['hostids'],
 			'pattern' => $filter['pattern'],
 			'time_from' => $time_from,
 			'time_to' => $time_to
@@ -119,9 +144,11 @@ class ReportsPercentiles extends CController {
 		$data = [
 			'filter' => $filter,
 			'groups' => $groups,
+			'hosts' => $hosts,
 			'active_tab' => CProfile::get(self::PROFILE_PREFIX.'active', 1),
 			'rows' => $rows,
 			'graph' => $graph,
+			'time_presets' => TimePresets::all(),
 			'definition' => $definition !== null
 				? ['reportid' => $definition['reportid'], 'name' => $definition['name']]
 				: null
